got rid of run on sentence, come back and complete later
Summary: This revision is a hotfix, I noticed that the end of the deployment docs had an incomplete phrase which was probably a typo in this PR: https://github.com/dagster-io/dagster/commit/c9bc657c6ed2cab2569120c68fe30879a88ed4de. This fixes it for posterity.

Test Plan: unit

Reviewers:

Subscribers: